Create hacktoberfestpraccepted
Open source is a term that originally referred to open source software (OSS). Open source software is code that is designed to be publicly accessible—anyone can see, modify, and distribute the code as they see fit. ... Open source has become a movement and a way of working that reaches beyond software production